+++ src/capteur_ir.h
@@ -2,7 +2,6 @@
 #define CAPTEUR_IR_H
 
 #include <LibRobus.h>
-#include "operations_generales.h"
 #include "conversions.h"
 
 #define INTERNE 0
@@ -13,8 +12,11 @@
 //Légèrement < distance entre les postes
 #define SEUIL_LETTRE_EXTERNE 20
 
-//Utilisée pour lire une valeur de distance en cm du capteur IR (index entre 0 et 3), en prenant amplitude*2-1 mesures
-float lireDistanceIR(uint8_t capteur, uint8_t amplitude = 3);
+//Prend un tableau d'int et sa longueur, et classe les membres du tableau en ordre croissant
+void classerTableauCroissant(uint16_t tableau[], const int NB_ELEMENTS);
+
+//Prend un tableau d'int et sa longueur (impaire), et renvoie la mediane parmi ses membres
+uint16_t calculerMediane(uint16_t tableau[], const int NB_ELEMENTS);
 
 //Utilisée pour lire directement une valeur entre 0 et 1023 du capteur IR (index entre 0 et 3), en prenant amplitude*2-1 mesures
 uint16_t lireValeurIR(uint8_t capteur, uint8_t amplitude = 5);
