+++ src/conversions.h
@@ -2,9 +2,6 @@
 #define CONVERSIONS_H
 #include <LibRobus.h>
 
-#include "capteur_couleur.h"
-#include "capteur_ir.h"
-
 extern const double CLICS_PAR_CM;
 //diametre de la roue en cm
 extern const float diaRoue;
@@ -16,11 +13,5 @@ double clicsEnCm(long);
 
 //Convertit une distance (en cm) en nombre de clics d'encodeur
 long long cmEnClics(float);
-
-//Convertit une entree analogique (0 à 1023) en tension (0.0 à 5.0 V)
-float entreeAnalogiqueEnTension(uint16_t entreeAnalogique);
-
-//Convertit une mesure de tension provenant d'un capteur infrarouge en distance en cm
-float tensionEnDistance(float tension);
 
 #endif
