+++ src/init_robot.h
@@ -16,10 +16,6 @@
 #include "suiveur_ligne.h"
 #include "servomoteur.h"
 
-#include "loop_normal.h"
-#include "loop_A.h"
-#include "loop_B.h"
-
 //Indice 0: où Bob-A va chercher le ballon
 //Indice 1: où Bob-B va porter le ballon
 extern const int COULEURS_BOB[];
