--- src/loop_A.h
@@ -1,11 +1,0 @@
-#ifndef LOOP_A
-#define LOOP_A
-
-#include "init_robot.h"
-
-void octogoneA(int couleur = -1500);
-void loopOctogoneA();
-void octogoneAlternatifA(int couleur = -1500);
-
-
-#endif
