--- src/loop_B.h
@@ -1,19 +1,0 @@
-#ifndef LOOP_B_H
-#define LOOP_B_H
-
-#include "init_robot.h"
-
-//Fonction loop utilisée par le défi du combattant, robot B
-//Prend en paramètre la couleur voulue, avec par défaut la valeur du tableau dans init_robot.cpp
-void loopOctogoneB(int COULEUR = -1500);
-
-//Fonction bloquante utilisée par le défi du combattant, robot bloquante
-//Prend en paramètre la couleur voulue, avec par défaut la valeur du tableau dans init_robot.cpp
-void octogoneB(int COULEUR = -1500);
-
-void octogoneAlternatifB(int COULEUR = -1500);
-
-
-
-
-#endif
