--- src/loop_normal.h
@@ -1,17 +1,0 @@
-#ifndef LOOP_NORMAL_H
-#define LOOP_NORMAL_H
-
-#include "init_robot.h"
-#include "interactions.h"
-#include "capteur_ir.h"
-#include "suiveur_ligne.h"
-#include "mouvement.h"
-
-//Loop standard
-void loopNormal();
-
-
-
-
-
-#endif
