+++ src/suiveur_ligne.cpp
@@ -184,105 +184,6 @@ void estLigneHuit(int lectures[8])
   }
 }
 
-void suivreLigneSimpleHuit(float DISTANCE, float VITESSE = 0.3)
-{
-  int huitLectures[8];
-  long long clicsTotaux = cmEnClics(DISTANCE);
-  float vitesseG = VITESSE;
-  float vitesseD = VITESSE;
-  int estFinLigne = 0;
-  int estPasLigneCentre = 0;
-  int estLigneCentre = 0;
-  int directionRotation = 0;
-
-  int CLICS_360_DEGRES = cmEnClics(largeurEss*2*PI);
-
-  const int DT = 5; //délais du cycle
-
-  bool ligneAuCentre = false;
-
-  resetDeuxEncodeurs();
-
-  while (ENCODER_Read(0)<clicsTotaux && estFinLigne<10)
-  {
-    lireSuiveurLigne(huitLectures);
-    estLigneHuit(huitLectures);
-
-    changerVitesseDeuxMoteurs(vitesseG, vitesseD);
-    delay(DT);
-
-    if(huitLectures[3]>=1 || huitLectures[4]>=1)
-    {
-      estLigneCentre++;
-    }
-    else
-    {
-      estPasLigneCentre ++;
-    }
-    if (estPasLigneCentre>30)
-    {
-      ligneAuCentre = false;
-      estPasLigneCentre = 0;
-    }
-    else if (estLigneCentre>30)
-    {
-      ligneAuCentre = true;
-      estLigneCentre = 0;
-    }
-
-    if (ligneAuCentre)
-    {
-      //Commenter ce if.. pour tourner en rond quand il voit pas de ligne, ce qui facilite leur recherche
-      if (huitLectures[CAPTEUR_DROIT]==huitLectures[CAPTEUR_GAUCHE])
-      {
-        vitesseG = VITESSE;
-        vitesseD = VITESSE;
-      }
-      if (huitLectures[CAPTEUR_DROIT] >= 1 && huitLectures[CAPTEUR_GAUCHE] >= 1)
-      {
-        estFinLigne++;
-      }
-      else if (huitLectures[CAPTEUR_GAUCHE]==1) //Capteur de gauche == ligne
-      {
-        // changerVitesseDeuxMoteurs(0, 0.2);
-        vitesseG = 0;
-        vitesseD = 0.2;
-        // vitesseD = vitesseD-0.02;
-        // changerVitesseMoteur(1, vitesseD);
-        // delay(500);
-      }
-      else if (huitLectures[CAPTEUR_DROIT]==1)   //Capteur de droite == ligne
-      {
-        // changerVitesseDeuxMoteurs(0.2, 0);
-        vitesseG = 0.2;
-        vitesseD = 0;
-        // vitesseD = vitesseD+0.02;
-        // changerVitesseMoteur(1, vitesseD);
-        // delay(500);
-      }
-    }
-    else
-    {
-      if (huitLectures[CAPTEUR_DROIT] >= 1 && directionRotation!=-1)
-      {
-        // changerVitesseDeuxMoteurs(-0.2, 0.2);
-        vitesseG = 0;
-        vitesseD = 0.2;
-        directionRotation = 1;
-      }
-      else if (huitLectures[CAPTEUR_GAUCHE] >= 1 && directionRotation!=1)
-      {
-        // changerVitesseDeuxMoteurs(0.2, -0.2);
-        vitesseG = 0.2;
-        vitesseD = 0;
-        directionRotation = -1;
-      }
-    }
-  }
-  arreterDeuxMoteurs();
-  resetDeuxEncodeurs();
-}
-
 void trouverLigne(int sensDeRotation)
 {
   int sens = (sensDeRotation==0) ? 1 : -1;
